requirements.txt
ultralytics==8.3.225
tensorflow==2.12.0
opencv-python
Pillow
matplotlib
fastapi
uvicorn
pytest
